feat: update distribusi and produksi metric datatype to numeric

+++ src/ddl/fact_produksi.sql
@@ -5,7 +5,7 @@ CREATE TABLE fact_produksi (
   id_unit_peternakan INT8,
   id_jenis_produk INT8,
   id_sumber_pasokan INT8,
-  jumlah_produksi INT8,
+  jumlah_produksi NUMERIC(10,3),
   created_dt TIMESTAMP,
   modified_dt TIMESTAMP,
   PRIMARY KEY(id_waktu, id_lokasi, id_unit_peternakan, id_jenis_produk, id_sumber_pasokan)
@@ -18,7 +18,7 @@ CREATE TABLE fact_produksi_stream (
   id_unit_peternakan INT8,
   id_jenis_produk INT8,
   id_sumber_pasokan INT8,
-  jumlah_produksi INT8,
+  jumlah_produksi NUMERIC(10,3),
   created_dt TIMESTAMP,
   modified_dt TIMESTAMP,
   PRIMARY KEY(id_waktu, id_lokasi, id_unit_peternakan, id_jenis_produk, id_sumber_pasokan)
